does not use custom package URLs in SetupDevEnv.jl after PR (#5)
After PR #3 I recognized, that lines starting with `CI_UNIT_PKG_URL_`
exist in the commit message of the merge request. This triggers
accidentally custom URLs for packages in the `SetupDevEnv.jl` script.
Because it does not make any sense to use custom URLs on the `main` and
`dev` branch, this PR disable the feature for this branches.

Add also gitlab CI attribute `interruptible: true` that CI can be
aborted if a new commit was pushed to the same branch.