use dev branches in integTestGen.jl to create dependency graph (#19)
Before the commit, integTestGen.jl uses the packages from the registry
to create the dependency graph. The causes a problem, if a QED sub
package adds a dependency to another QED sub package and this change was
not released. In this case, the generator takes the old version of the
registry without the new dependency and does not generate a integration
test for the package. Now it uses the dev branch by default or a custom
repository and branch defined by the environment variable
`CI_INTG_PKG_URL_*`.

Co-authored-by: Anton Reinhard <s1509337@msx.tu-dresden.de>