fix integration test for merge in the main branch (#63)
If we want to merge in the main branch, we do it because we want to
publish the package. Therefore, we need to be sure that there is an
existing version of the dependent QED packages that works with the new
version of the package we want to release. The integration tests are
tested against the development branch and the release version.
- The dev branch version must pass, as this means that the latest
version of the other QED packages is compatible with our release
version.
- The release version can exist. If they exist, it means that we do not
need to release a new version of this package.

I will open an test PR to test if the feature is correctly working. If
it is working, I mark the PR for ready for review.