Implement CI bootloader script (#86)
At the moment, sometimes if I changes something on the CI, I also need
to modify the content of `.gitlab-ci.yaml` in the projects, which uses
the CI scripts. At the moment, there are 4 projects, where I need to do
the same changes. To remove the need of this redundant work, I want to
implement a CI bootloader.

The basic idea of the bootloader is, there are only two jobs defined in
the `.gitlab-ci.yaml`. The first job downloads the CI bootloader and
execute it. The bootloader, scans it's environment (environment
variables) and creates depending on it a GitLab CI yaml file, which
contains the whole test pipeline such as unit and integration tests. The
second job simply executes the CI pipeline.

In the best cases, we don't need to change the `.gitlab-ci.yaml`
anymore, if we change the CI scripts. Maybe it is required to set
optional variables for single projects. For example we change the
default Julia version for integration tests from 1.10 to 1.11 but
project X does not support it yet. Therefore it still needs to run with
version 1.10 until the problem is solved.

Possible `.gitlab-ci.yaml`:

```yaml
stages:
- generate
- run

generate_pipeline:
  image: julia:1.10
  stage: generate
  script:
    - apt update && apt install -y git
    - git clone --depth 1 -b dev https://github.com/QEDjl-project/QuantumElectrodynamics.jl.git /generator
    - julia --project=/generator/.ci/CI -e 'import Pkg; Pkg.instantiate()'
    - julia --project=/generator/.ci/CI bootloader.jl > $CI_PROJECT_DIR/pipeline.yaml
  artifacts:
    paths:
      - pipeline.yaml
    expire_in: 1 week
  interruptible: true
  tags:
    - cpuonly

run_pipeline:
  stage: run
  trigger:
    include:
      - artifact: pipeline.yaml
        job: generate_pipeline
    strategy: depend
 ```