CI Bootloader.jl: allow to write unit test environment variable verify job in extra yaml file and run in extra pipeline (#116)
Allow to write unit test environment variable verify job in extra yaml
file and run in extra pipeline.

fix: #112 

# CI Test

- test PR: https://github.com/QEDjl-project/QEDprocesses.jl/pull/125
- Pipeline with unmodified `.gitlab-ci.yaml`:
https://gitlab.com/hzdr/qedjl-project/QEDprocesses-jl/-/pipelines/1781442408

# Warning

If the PR is merged and the `.gitlab-ci.yaml` of a project that uses
`Bootloader.jl` is not modified, the verification job will not be
executed. The rest of the tests will work as expected.