Bootloader.jl write dummy jobs to empty file output (#122)
- If the user specify that jobs should be written to file but no jobs
are generated for the output, generate dummy job and write it file.
- The file output is used for GitLab CI child pipelines. If the files
are empty, the pipeline will fail.
- The use case is the following. Output to files needs to be static
configured in the .gitlab-ci.yaml because the child pipelines needs to
be static configured. In future it should be possible to enable or
disable generating specific CI jobs via PR comments and/or PR tags.
Therefore it is possible that no CI jobs are generated for a specific
file output.