use QEDCore package to verify Bootloader.jl code in the CI (#125)
Triggers automatically the CI of the QEDCore.jl project with modified
code of Bootloader.jl as part of the CI in a pull request.
Replace the need to create test PRs in other projects, like this PR:
https://github.com/QEDjl-project/QEDcore.jl/pull/114

Note: I tested with a direct push to GitLab.com that the job is note
executed if we are not in GitHub pull request. Still there is a chance,
that I did something wrong. I use the feature for the fist time. So we
will see if it really works after merging.