moved tests from bootloader_integration_tests.yaml to Julia tests (#128)
I realized that it is pretty hard to run the `Bootloader.jl` integration
tests during the work on PR #127. Also it is pretty easy to miss the
tests until the CI crashes.

- ~[ ] try to parallelize `Bootloader.jl` integration tests~

It does not make sense to parallelize. I improved the test runtime time
from 51s to 46s on my laptop with 4 P-Cores. I think the reason is, that
it cannot cache the jited code. Therefore it compiles the code of
`Bootloader.jl` 4 times and then execute the actual test.
Therefore, I got some restrictions to write test code.