get_gitlab_ci_conf.jl: allow to overwrite specific environment variables (#138)
In PR #137 I implemented a function, which checks if a environment
variable is already defined and does not overwrite it's value. But this
behavior is not correct for each environment variable. For example the
environment variable `CI_QED_ENABLE_CUDA_TESTS` should be overwritten if
a specific label is set. This PR fix this.

Add also additional logger output. If `$(julia get_gitlab_ci_conf.jl)`
is executed, it is not visible which environment variables are set.
Therefore display this via info logger.